pthreads: condattr_t and mutexattr_t functions

+++ libs/libc/include/bits/clockid_t.h
@@ -5,7 +5,7 @@
 extern "C" {
 #endif /* __cplusplus */
 
-typedef unsigned long clockid_t;
+typedef int clockid_t;
 
 #ifdef __cplusplus
 }
